+++ packages/base/sync/__tests__/applyChanges.spec.ts
@@ -18,40 +18,40 @@ function getDefaultChangeItem(id = '1') {
   return { id, time: Date.now(), collectionName: 'test' }
 }
 
-it('should insert new items', () => {
+it('should insert new items', async () => {
   const items: TestItem[] = [{ id: 1, name: 'Item 1' }]
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem(), type: 'insert', data: { id: 2, name: 'Item 2' } },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([
     { id: 1, name: 'Item 1' },
     { id: 2, name: 'Item 2' },
   ])
 })
 
-it('should update an item', () => {
+it('should update an item', async () => {
   const items: TestItem[] = [{ id: 1, name: 'Item 1' }]
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem(), type: 'update', data: { id: 1, modifier: { $set: { name: 'Updated Item 1' } } } },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([{ id: 1, name: 'Updated Item 1' }])
 })
 
-it('should remove an item', () => {
+it('should remove an item', async () => {
   const items: TestItem[] = [{ id: 1, name: 'Item 1' }, { id: 2, name: 'Item 2' }]
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem(), type: 'remove', data: 1 },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([{ id: 2, name: 'Item 2' }])
 })
 
-it('should handle a mix of inserts, updates, and removes', () => {
+it('should handle a mix of inserts, updates, and removes', async () => {
   const items: TestItem[] = [{ id: 1, name: 'Item 1' }]
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem('1'), type: 'insert', data: { id: 2, name: 'Item 2' } },
@@ -59,28 +59,28 @@ it('should handle a mix of inserts, updates, and removes', () => {
     { ...getDefaultChangeItem('3'), type: 'remove', data: 2 },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([{ id: 1, name: 'Updated Item 1' }])
 })
 
-it('should update already existing items with insert', () => {
+it('should update already existing items with insert', async () => {
   const items: TestItem[] = [{ id: 1, name: 'Item 1' }]
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem(), type: 'insert', data: { id: 1, name: 'New Item Name' } },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([
     { id: 1, name: 'New Item Name' },
   ])
 })
 
-it('should insert non-existing items on update', () => {
+it('should insert non-existing items on update', async () => {
   const items: TestItem[] = []
   const changes: Change<TestItem, number>[] = [
     { ...getDefaultChangeItem(), type: 'update', data: { id: 1, modifier: { $set: { name: 'Updated Item 1' } } } },
   ]
 
-  const result = applyChanges(items, changes)
+  const result = await applyChanges(items, changes)
   expect(result).toEqual([{ id: 1, name: 'Updated Item 1' }])
 })